In Update Project example, do not set admin_ids
Due to a bug, admin_ids or group_admin_ids must be supplied when
performing an update. As a result, the prior version of this script
sets the current user as an admin and this necessitates warning text.

Remove that setting to avoid having a warning which is based around a
bug. This means that the example cannot be considered ready to release
until the bug is resolved.