Add methods for index {create,delete,reopen}
These are a trio of Search API methods. Here they are introduced
without test data, tests, or anything beyond their basic call
structure and docs.

Because the docs reference fixture data which does not yet exist, the
docs are not fully valid yet.

This change includes a changelog for the new methods.