Allow base_url without service_name in clients
base_url can now be used to effectively skip the check for
`service_name == "_base"`, which has two significant impacts.

The first is simply that it is now possible to instantiate a
BaseClient directly, which may prove useful for ad-hoc scripting.
Although not generally a recommended approach (e.g. there is no
benefit to using a BaseClient in lieu of a TransferClient), the
possibility only makes the SDK more flexible.

The second effect is that a class inheriting from BaseClient no longer
needs to set `service_name` if the setting has no appropriate value.
This is relevant for 3rd party clients with nonstandard URL schemes
(e.g. FuncX before the Compute rebranding) and for clients to an array
of potential hosts, like GCSClient and any future
ActionProviderClient.

The only new test ensures that it is possible to instantiate a
BaseClient directly, and the GCSClient.service_name value, although
unchanged, is marked for a v4.0 update.