Fix typing bug in PaginatorTable declarations
A `Paginator[T]` when `T` is unbound is effectively the same as
`Paginator[Any]` but does not allow any other use of the same unbound
`T` to match. It's unclear if this should be allowed -- in fact, the
use of the unbound type var is the main issue here.
To resolve, switch the unbound type var to `Any`.