Improve SpecificFlowClient.scopes typing
`mypy` would flag the `scopes` object as `ScopeBuilder | None` even
though the class always defined a de-facto value for this attribute on
init.

To resolve, insert a stub object as a class variable. Not only does
this pacify mypy, but it also allows us to raise more useful errors
when the classvar is accessed.
Define the stub and several tests for it, including a typing test.