utils.classproperty now prefers the object
The behavior of preferring an instance over a class, although valid
and normal for a descriptor, is unlike classmethod and therefore may
be surprising. However, it better allows us to define differentiated
behaviors between instances and their classes, especially with respect
to the `scopes` object on client classes.

Without this change, it is not possible for a classproperty, when
called, to refer back to the instance on which it was invoked.
However, it is always possible to find the class via the instance, so
expressive power is only increased by this change.

Institute a convention of the classproperty decorated method accepting
`self_or_cls` to clarify this at the usage sites.

Comments on `classproperty` are updated. In particular, make sure that
the current status of `classmethod(property(...))` stacking is
reflected in those comments.

A new test validates that the class attribute remains valid but the
instance attribute is preferred when a classproperty is used to access
an attribute.