Minor change to stub scopes object to pass pylint
pylint complains about an `__init__` not invoking `super().__init__`.
If we do the "obvious" thing and call the superclass init with a
placeholder value, attribute access patterns change.

In particular, `self.resource_server` is assigned as an attribute,
meaning that the concrete value there is discovered by the default
`__getattribute__` before `__getattr__` can be invoked. To solve this,
intercept that access with a custom `__getattribute__` to raise an
error. The error will implicitly be handled and trigger a call of
`__getattr__`, resulting in the desired `AttributeError` for
`resource_server` access.