Implement _guards.reduce for better chaining
Validators in the experimental GARE subpackage can be built from
TypeGuards if those type guards can be built in a form which is
suitable to some kind of chaining API in which the type is an input
and the guards are outputs or attributes.

`reduce` is a purpose-built helper in `_guards` which provides a
`_Reducer[T]` if given a `type[T]`.
`reduce(str) -> _Reducer[str]`.

The `_Reducer` can then have methods which act as guards of the
desirable signatures.

Finally, this can be pulled into the GARE component to write a
TypeGuard -> ValidatorCallback helper and wrap the results of `reduce`
calls.