Use `all` in default_scope_requirements for Flows
Update the `default_scope_requirements` field for `FlowsClient` to
list (only) the `all` scope. This increases the capabilities of a
default-built FlowsClient to include creation of a Flow.

This appears to have been a minor oversight in the initial
`default_scope_requirements` build-out, not an intentional decision to
limit capabilities. The `all` scope is newer than
`default_scope_requirements` and is the most appropriate choice now.