Rename Timers classes and alias old names
The following renames are performed:

    globus_sdk.TimerClient => globus_sdk.TimersClient
    globus_sdk.TimerAPIError => globus_sdk.TimersAPIError
    globus_sdk.scopes.TimerScopes => globus_sdk.scopes.TimersScopes

For each of these names, the previous name is retained as an alias.
In the case of the 'TimerScopes' object, a deprecation warning is
emitted if the old name is used -- encouraging users to move to
'TimersScopes' whenever possible.

For 'TimerClient' and 'TimerAPIError', no deprecation warning is
emitted yet -- but code for such a warning is suggested in a comment.
The purpose behind this is to have one or more releases in which both
names are supported before beginning to warn on use of the old name.
This gentle deprecation path is not offered for the scopes object due
to its more niche usage -- and the fact that it can be alternatively
accessed via the client object.

Numerous doc references are updated, and some tests are added to check
the aliases. Tests are not converted here to use the new names.