Promote `gare` from `experimental`
The `auth_requirements_error` module is moved to `globus_sdk.gare` and
docs and importts are updated.

Included in this change, `GlobusAuthRequirementsError` has been
renamed to `GARE`, as the `Error` suffix to a name typically indicates
that a class is an exception class.

Aliasing has been put in place from `experimental` to the new code
location.

As a minor, related fix `_serializable.Serializable` now uses a
`Self` type annotation, rather than a type var. This improves the doc
rendering of the `from_dict` and `to_dict` methods, as they show the
self-type rather than an unbound type variable T.