Control determination of `identity_id` by context
Depending on the response context (the type of OAuthTokenResponse in
use) and the calling context (ValidatingTokenStorage or the
base class), there are various optimal behaviors for `identity_id`
determination.

In order to avoid declaring any new public interface which would need
support, this is defined as an internal-only hook,
`TokenStorage._extract_identity_id`, and given various behaviors in
different scenarios.

In the typical case, it acts exactly as it used to. The following
scenarios modify its behavior:
- on dependent token responses, it returns None
- in the ValidatingTokenStorage, it returns
  `ValidatingTokenStorage.identity_id` IFF the response is an
  OAuthRefreshTokenResponse