Test tokenstorage handling of various types
Define tests which ensure that all of the standard TokenStorage
classes handle the variants of OAuthTokenResponse gracefully.
Refresh and dependent token responses should be accepted, in spite of
their data shapes not including `id_token`.

In the case of refresh token data, `id_token` is absent, and in the
case of dependent token data, the top-level structure is an array.

As a control against these two tests, check the behavior of an
authorization code grant response containing an `id_token`.