Further refactor ValidatingTokenStorage
Move the implementation into a nested subpackage and introduce a
TokenValidationContext object which allows the storage to pass state
down to the validators in a structured manner.

Validators are now declared as a protocol (and there is only one such
protocol), and both before_store and after_retrieve are coerced into
the same data-shape for simplicity. This also allows for batched
execution of the after_retrieve validators, when getting all token
data from storage.

Names and export symbols are updated as appropriate, and tests require
only minimal tweaks. All of the relevant elements of this system are
now provided as `globus_app`-level exports.

Internal validators are declared to encode the
AccessTokenAuthorizerFactory and RefreshTokenAuthorizerFactory checks,
but the structure by which these are applied is not yet changed.