Remove literal type alias to help sphinx render
This type alias is used in 3 locations (two of which are internal),
but sphinx autodoc does not render it as desired or expected. It seems
that there's something more complex going wrong with the docstring for
this class, but it may also be related to some outstanding sphinx bugs
related to type aliases and literals, specifically.

For now, switch off of the alias and use the literal verbatim in
`TransferData.__init__` to get a rendering which is much closer to
being correct.