Extract and re-inject base client sphinx param doc
Add three helpers to `utils` in order to improve the runtime
docstrings for client classes.

1. `read_sphinx_params` -- a simple `:param:` parser
2. `inject_sphinx_params` -- a `__doc__` rewriter which takes parsed
        params and inserts them in place of a special marker comment
3. `inject_sphinx_params_of` -- a decorator version of these which
        chains together reading the docstring of one object and
        modifying the docstring of another

Then apply these to write

    ```python
    @utils.inject_sphinx_params_of(client.BaseClient)
    class FlowsClient(client.BaseClient):
        ...
    ```

and so forth. This is applied to all client classes except for the
Auth login clients, which do not inherit all of the parameters of
`BaseClient` -- in particular, we do not allow construction of a
login client with an `app`.

`read_sphinx_params` gets a functools cache, but the rest of the work
is still done dynamically, so there is a little bit of repetition.

This results in improved docstrings for sphinx and for runtime usages
like `help(globus_sdk.TransferClient)`. Static analyzers (e.g., IDEs)
will pick up on the original source string, which will show the marker
comment used to drive this behavior:

    ```
    .. globus-sdk-inject-doc-params
    ```