Fix the presentation of scope constants in docs
At some point, sphinx autodata directives for autodoc changed in
behavior. autodoc stopped picking up on the object docstrings we had
alongside each constant and started using the class docstring for each
element, which is often inappropriate to the relevant item.

Moving to explicit `py:data` directives evades any such issue and
gives us an opportunity to change the placement of admonitions (note
and warning usage) to be part of the data docstrings.