Better support parsing a GARE from the error root
When parsing a GARE from an API error object, we previously would
consider the sub-errors in `err.errors` (where `err` is a
`GlobusAPIError`) but would not inspect the root of the error object.
We now parse a GARE from the root error in these situations.

Because the default parse for non-JSON:API errors populates
`err.errors` with `data["errors"]` if present, *and falls back* to
`[data]` otherwise, there's some ambiguity we need to resolve. It's
important that `to_gares`, the plural/list variant, does not create
two GAREs when there is only one, found at the root.

To make the revised code simple, the conversion of
`dict | None -> GARE | None` is separated from the main `to_gare`
interface in a minor refactor.