Convert the parser perf test to pytest-benchmark
This is an experiment in the use of `pytest-benchmark` to provide a
structure for benchmarking in general. The perf test is moved and
rephrased as a set of pytest test cases using the `benchmark` fixture.

A new testenv, `tox r -e benchmark`, is defined to allow runs of these
benchmarks.

The `tests/benchmark/` dir is listed in norecursedirs, meaning that we
don't need `pytest-benchmark` installed at all outside of the
benchmark testing tox testenv.