Change how `BaseClient.transport` is customized
Allow users to pass in a constructed transport, rather than always
requiring that they pass parameters which the client will then use to
instantiate a new transport.

This enables experimentation with transport reuse, sharing, and other
paradigms of use which were previously not naturally possible.

It also reduces a typing-time imprecision in the client initialization
API, where the transport parameters are expected to match the
transport class, but this cannot be enforced contractually because the
class itself is dynamic.

Rather than providing a `transport_class`, clients now have a
`default_transport_factory`. This is typically going to be a class,
but conceptually it's any callable with no arguments.