Introduce RetryCheckCollection
This replaces transport object factories. Instead, each client simply
- has a transport (which the user may pass)
- has a retry check collection (which is user-accessible but may not
  be passed on init)

The retry check collection is the list-of-checks part of the
transport, abstracted out into a separate type. It is passed into the
transport as part of the caller info.

This simplifies the transport itself, and makes each client
independently responsible for its retry checks. As a result, the
Transfer client can use the base transport type without issue.