Refactor RetryConfig to isolate hook registration
In order to give TransferClient control over its retry checks but
still have the retry config exposed to user control, hook registration
is moved off of the core models for retry config and instead now lives
in a callback in the client definitions. The base client registers the
default hooks (which are just provided as a tuple).

The TransferClient registers its own modified suite of hooks.

Additionally:
- rename to 'retry_config / RetryConfig' to match 'GlobusAppConfig'
- clean up some tests to use simpler call styles
- add some missing functional tests of transfer retries
- update the RetryConfig docs