Improve structure of resource closure in GlobusApp
In order to accurately track resources and `close()` when appropriate, a
new list tracks the resources to close. The client is adapted to this
same pattern for consistency.

Both objects log debug level messages when closing to indicate what is
being closed.

Some indirection is still needed so that the creation flow for
`GlobusApp._token_storage` can determine whether or not
`GlobusApp.token_storage` will be closed.