Add a doc to the User Guide for combining GAREs
This document covers, in brief, why combining GAREs is both desirable and
not safe in many circumstances. It also offers an example of a merge
algorithm which can be used in limited cases to simplify a list of GAREs
into a smaller list, including a "full demo" run in the `__main__` block.